+++ src/app/locales/en/components/cluster-storage-add-form.js
@@ -2,7 +2,6 @@ import generic from './cluster-storage-add-form/generic';
 import luma from './cluster-storage-add-form/luma';
 import ceph from './cluster-storage-add-form/ceph';
 import cephrados from './cluster-storage-add-form/cephrados';
-import embeddedceph from './cluster-storage-add-form/embeddedceph';
 import posix from './cluster-storage-add-form/posix';
 import s3 from './cluster-storage-add-form/s3';
 import swift from './cluster-storage-add-form/swift';
@@ -18,7 +17,6 @@ export default {
   luma,
   ceph,
   cephrados,
-  embeddedceph,
   posix,
   s3,
   swift,