+++ src/app/locales/en/components/content-clusters-ceph.js
@@ -1,7 +1,0 @@
-export default {
-  header: 'Ceph',
-  status: 'Status',
-  configuration: 'Configuration',
-  pools: 'Pools',
-  cephNotDeployed: 'Ceph not deployed',
-};