+++ src/app/locales/en/components/sidebar-clusters.js
@@ -2,7 +2,6 @@ export default {
   secondLevelItems: {
     dnsWarning: 'Some problems with your DNS configuration have been detected',
     webCertWarning: 'Some problems with your web certificate have been detected',
-    cephWarning: 'Some problems with your Ceph cluster have been detected',
     noConnectedUserWarning: 'This cluster does not have any linked Onezone user',
     guiSettingsWarning: 'Some problems with your GUI settings have been detected',
   },